+++ src/core/primitives/RovingFocusGroup/RovingFocus.tsx
@@ -1,0 +1,10 @@
+
+import RovingFocusGroup from "./fragments/RovingFocusGroup"
+import RovingFocusItem from './fragments/RovingFocusItem';
+
+const RovingFocus = {
+    Group : RovingFocusGroup,
+    Item : RovingFocusItem
+} as const 
+
+export default RovingFocus;
